requirements.txt
Flask==3.1.2
openai==1.0.0
requests==2.32.5
python-dotenv==1.2.1
numpy==2.3.5
scikit-learn==1.7.2
tiktoken==0.12.0
